fix: fix and test remove and find client and employee use cases

+++ src/main/java/com/example/sistemacabeleleiro/Domain/UseCases/Employee/RemoveEmployeeUseCase.java
@@ -1,5 +1,6 @@
 package com.example.sistemacabeleleiro.Domain.UseCases.Employee;
 
+import com.example.sistemacabeleleiro.Domain.Entities.Client.Client;
 import com.example.sistemacabeleleiro.Domain.Entities.Employee.Employee;
 import com.example.sistemacabeleleiro.Domain.Entities.Employee.EmployeeStatus;
 import com.example.sistemacabeleleiro.Domain.Entities.Schedulling.Scheduling;
@@ -18,21 +19,27 @@ public class RemoveEmployeeUseCase {
     }
     public boolean remove(Integer id){
         var employee = employeeDAO.findOne(id);
+
         if (id == null || employee.isEmpty()){
             throw new EntityNotFoundException("Employee not found");
         }
-
+        if (!employeeSchedules(employee.get()).isEmpty()){
+            throw new IllegalArgumentException("It's not possible to exclude employees who already have a schedule");
+        }
         validateEmployeeStatus(employee.get());
-        deleteEmployeeSchedules(employee.get());
+
         return employeeDAO.deleteByKey(id);
     }
     public boolean remove(Employee employee){
         if (employee == null || employeeDAO.findOne(employee.getId()).isEmpty()){
             throw new EntityNotFoundException("Employee not found");
         }
+        if (!employeeSchedules(employee).isEmpty()){
+            throw new IllegalArgumentException("It's not possible to exclude employees who already have a schedule");
+        }
 
         validateEmployeeStatus(employee);
-        deleteEmployeeSchedules(employee);
+
         return employeeDAO.delete(employee);
     }
 
@@ -42,14 +49,11 @@ public class RemoveEmployeeUseCase {
         }
     }
 
-    private void deleteEmployeeSchedules(Employee employee) {
-        List<Scheduling> schedulesToDelete = schedulingDAO.findAll().stream()
+    private List<Scheduling> employeeSchedules(Employee employee) {
+        List<Scheduling> schedules = schedulingDAO.findAll().stream()
                 .filter(schedule -> schedule.getEmployee().equals(employee))
                 .toList();
-        if (!schedulesToDelete.isEmpty()){
-            for (Scheduling schedule : schedulesToDelete) {
-                schedulingDAO.delete(schedule);
-            }
-        }
+
+        return schedules;
     }
 }
